requirements.txt
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
deepspeech==0.9.3
deepspeech-tflite==0.9.3
enum34==1.1.10
idna==2.10
numpy==1.21.0
pkg-resources==0.0.0
pvporcupine==1.9.5
py-espeak-ng==0.1.8
PyAudio==0.2.11
pycairo==1.20.1
pycparser==2.20
PyGObject==3.40.1
requests==2.25.1
scipy==1.7.0
SoundFile==0.10.3.post1
urllib3==1.26.5
webrtcvad==2.0.10
